Move all function component defination outside of the Gameroom. fix #89
